Merge "about" and "README"
Since there is only one file in this folder, it makes sense to merge it with the README. That way, users don't have to click a second time to get to the content. if you like this change, you can accept the pull request and merge it back into your project.